backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
pydantic==2.5.0
python-decouple==3.8
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
gunicorn==21.2.0
geopy==2.4.0
folium==0.15.0
plotly==5.17.0 